+++ R/multicore.R
@@ -17,9 +17,13 @@
 #' plan(multicore, workers = 2)
 #' ```
 #'
-#' @inheritParams multiprocess
-#' @inheritParams Future-class
 #' @inheritParams future
+#' @inheritParams Future-class
+#' @inheritParams MulticoreFuture-class
+#'
+#' @param workers The number of parallel processes to use.
+#' If a function, it is called without arguments _when the future
+#' is created_ and its value is used to configure the workers.
 #'
 #' @return
 #' A \link{MulticoreFuture}.