+++ R/multisession.R
@@ -18,10 +18,12 @@
 #' plan(multisession, workers = 2)
 #' ```
 #'
-#' @inheritParams multiprocess
+#' @inheritParams multicore
 #' @inheritParams cluster
 #' @inheritParams Future-class
 #' @inheritParams future
+#' 
+#' @param \dots Additional arguments passed to [Future()].
 #'
 #' @param rscript_libs A character vector of \R package library folders that
 #' the workers should use.  The default is `.libPaths()` so that multisession